+++ src/FileIO/FIFileBinary.cpp
@@ -69,7 +69,7 @@ void FIFileBinary::LoadBlock(){
 
 		// 可変長フィールドの場合，長さを読み込みセット
 		if(field->varType == UTTypeDesc::Field::VECTOR){
-			fieldIt.arrayLength = *(size_t*)ptr;
+			fieldIt.arrayLength = (int) *(size_t*)ptr;
 			ptr += sizeof(size_t);
 		}
