+++ src/Foundation/UTTypeDesc.cpp
@@ -65,7 +65,7 @@ void* UTTypeDesc::Field::GetAddressEx(void* base, int pos){
 		while((int)type->VectorSize(ptr)<=pos) type->VectorPush(ptr);
 		if (pos == -2){
 			type->VectorPush(ptr);
-			int back = type->VectorSize(ptr)-1;
+			int back = (int)type->VectorSize(ptr)-1;
 			ptr = type->VectorAt(ptr, back);
 			//	DSTR << type->GetTypeName() << "ArrayPos of -2, get ptr of " << back << std::endl;
 		}else{
