+++ src/HumanInterface/HISpidarCalc.cpp
@@ -13,7 +13,7 @@
 namespace Spr{;
 
 void HISpidarCalcBase::Init(int dof, const VVector<float>& minF, const VVector<float>& maxF){
-	int wire = minF.size();
+	size_t wire = minF.size();
 	matA.resize(wire, dof);
 	lengthDiff.resize(wire);
 	lengthDiffAve.resize(wire);
@@ -47,7 +47,7 @@ void HISpidarCalc3Dof::Init(int dof, const VVector<float>& minF, const VVector<f
 void HISpidarCalc3Dof::SetForce(const Vec3f& TF){
 	int i,j;
 	
-	int wire = wireDirection.size();
+	size_t wire = wireDirection.size();
 	for(i=0;i<wire;i++){
 		quadpr.matQ[i][i] = wireDirection[i]*wireDirection[i]+smooth;
 		for(j=i+1;j<wire;j++){
@@ -67,7 +67,7 @@ void HISpidarCalc3Dof::Update(){
 
 	MeasureWire();
 	MakeWireVec();
-	int wire = matA.height();
+	size_t wire = matA.height();
 	for(int step=0; step< nRepeat; ++ step){
 		for(i=0;i<wire;i++) {
 			matA[i][0]=wireDirection[i].X();
@@ -125,7 +125,7 @@ void HISpidarCalc6Dof::Update(){
 	MeasureWire();	//get length from encorder
 	MakeWireVec();
 
-	int nWire = matA.height();
+	size_t nWire = matA.height();
 	for(int step=0; step < nRepeat; ++step){
 		for(i=0;i<nWire;i++) {
 			matA[i][0]=wireDirection[i].X();
@@ -176,7 +176,7 @@ void HISpidarCalc6Dof::Update(){
 }
 
 void HISpidarCalc6Dof::SetForce(const Vec3f& trnF ,const Vec3f& rotF){
-	int nWire = wireDirection.size();
+	size_t nWire = wireDirection.size();
 
 	int i,j;
